[FIX] web: fix public env by introducing common env
A lot of keys were missing and services were not instantiated as expected.
